+++ CHANGELOG.md
@@ -117,6 +117,20 @@ Those changes in added, changed or breaking changes, should include usage exampl
     .status(); 
   // tx_status should be true
   ```
+<<<<<<< HEAD
+
+* Added `get_operator_restaked_strategies` in `avsregistry/reader` in [#348](https://github.com/Layr-Labs/eigensdk-rs/pull/348).
+
+  ```rust
+    let strategies = avs_reader
+      .get_operator_restaked_strategies(FIRST_ADDRESS)
+      .await
+      .unwrap();
+  ```
+
+* Added custom configuration for release-plz in [#281](https://github.com/Layr-Labs/eigensdk-rs/pull/281).
+=======
+>>>>>>> dev
 * Added Rewards2.1 support in [#323](https://github.com/Layr-Labs/eigensdk-rs/pull/323).
 
   * Set an operator's split on an operator set.
@@ -151,7 +165,7 @@ Those changes in added, changed or breaking changes, should include usage exampl
 * Added new method `set_operator_set_param` in `avsregistry/writer` in [#327](https://github.com/Layr-Labs/eigensdk-rs/pull/327).
 
   ```rust
-   let operator_set_params = OperatorSetParam {
+    let operator_set_params = OperatorSetParam {
         maxOperatorCount: 10,
         kickBIPsOfOperatorStake: 50,
         kickBIPsOfTotalStake: 50,
@@ -256,6 +270,47 @@ Those changes in added, changed or breaking changes, should include usage exampl
         .set_account_identifier(new_identifier_address)
         .await
         .unwrap();
+  ```
+* Added missing StakeRegistry writer functions in [#343](https://github.com/Layr-Labs/eigensdk-rs/pull/343).
+
+  * `set_minimum_stake_for_quorum`
+
+    ```rust
+    let tx_hash = avs_writer
+      .set_minimum_stake_for_quorum(quorum_number, minimum_stake)
+      .await
+      .unwrap();
+    ```
+
+  * add_strategies
+
+  ```rust
+  let tx_hash = avs_writer
+    .add_strategies(quorum_number, vec_of_strategy_params)
+    .await
+    .unwrap();
+  ```
+
+  * remove_strategies
+
+  ```rust
+  let tx_hash = avs_writer
+    .remove_strategies(quorum_number, indices_to_remove)
+    .await
+    .unwrap();
+  ```
+
+  * modify_strategy_params
+
+  ```rust
+  let tx_hash = avs_writer
+      .modify_strategy_params(
+          quorum_numbers,
+          vec_of_strategy_indices,
+          vec_of_new_multipliers,
+      )
+      .await
+      .unwrap();
   ```
 * Added missing stake registry view methods in `avsregistry/reader` in [#347](https://github.com/Layr-Labs/eigensdk-rs/pull/347).